pyproject.toml
[tool.poetry]
name = "aichat"
version = "0.1.0"
description = ""
authors = ["olegvallas <oleg.vallas@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.98.0"
uvicorn = "^0.22.0"
sqlalchemy = "^2.0.17"
asyncpg = "^0.27.0"
alembic = "^1.11.1"
sqlalchemy-utils = "^0.41.1"
loguru = "^0.7.0"
greenlet = "^2.0.2"
openai = "^0.27.8"
python-dotenv = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
